requirements.txt
pandas
numpy
ollama-client
chromadb
networkx
matplotlib
seaborn
scikit-learn
click
lxml
